Feature/via coverage

+++ lib/coverband/adapters/file_store.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 module Coverband
   module Adapters
-    class FileStore
+    class FileStore < Base
       attr_accessor :path
 
-      def initialize(path, opts = {})
+      def initialize(path, _opts = {})
         @path = path
 
         config_dir = File.dirname(@path)
@@ -11,18 +13,16 @@ module Coverband
       end
 
       def clear!
-        if File.exist?(path)
-          File.delete(path)
-        end
+        File.delete(path) if File.exist?(path)
       end
 
       def save_report(report)
         results = existing_data(path)
         report.each_pair do |file, values|
-          if results.has_key?(file)
+          if results.key?(file)
             # convert the keys to "3" opposed to 3
             values = JSON.parse(values.to_json)
-            results[file].merge!( values ){|k, old_v, new_v| old_v.to_i + new_v.to_i}
+            results[file].merge!(values) { |_k, old_v, new_v| old_v.to_i + new_v.to_i }
           else
             results[file] = values
           end
@@ -53,7 +53,6 @@ module Coverband
           {}
         end
       end
-
     end
   end
 end
